backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.5.1
pymongo==4.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
pydantic[email]==2.9.1
pydantic-settings==2.5.2
slowapi==0.1.9
python-multipart==0.0.9
certifi==2024.8.30
